reverted back to working release of skimmer to push to master
