+++ KUCMSSkimmer/makefile
@@ -16,7 +16,8 @@ RFGLIBS     = $(shell restframes-config --libs)
 BCFLAGS += -I/uscms/home/jaking/nobackup/el8/llpana/CMSSW_13_3_3/src/BayesianClustering/include -frounding-math
 BCGLIBS += -L/uscms/home/jaking/nobackup/el8/llpana/CMSSW_13_3_3/src/BayesianClustering/lib -lBayesianClusteringi
 
-#CGFLAGS += -I/uscms/home/jaking/nobackup/el8/llpana/CMSSW_13_3_3/src/
+HHFLAGS += -I/uscms/home/jaking/nobackup/el8/llpana/CMSSW_13_3_3/src/KUCMSNtupleizer/KUCMSNtupleizer/interface
+
 CGFLAGS += -I/uscms/home/jaking/nobackup/el8/llpana/CMSSW_13_3_3/src/CGAL-5.6.1/include
 FDFLAGS += -I/uscms/home/jaking/nobackup/el8/llpana/CMSSW_13_3_3/src/frugally-deep-master/include
 EGFLAGS += -I/uscms/home/jaking/nobackup/el8/llpana/CMSSW_13_3_3/src/eigen
@@ -30,7 +31,7 @@ p2=spaceholder
 all: $(p1) 
 
 $(p1): $(p1).cpp
-	$(CC) -o $(p1).obj $(p1).cpp $(RTCFLAGS) $(RTGLIBS) $(RFCFLAGS) $(RFGLIBS) $(BCFLAGS) $(BCGLIBS) $(FDFLAGS) $(EGFLAGS) $(FPFLAGS) $(CGFLAGS)
+	$(CC) -o $(p1).obj $(p1).cpp $(RTCFLAGS) $(RTGLIBS) $(RFCFLAGS) $(RFGLIBS)  
 
 $(p2): $(p2).cpp
 	$(CC) -o $(p2).obj $(p2).cpp $(RTCFLAGS) $(RTGLIBS)
